Extract classes from people editor files
