+++ src/gui/desktop/utils/CMakeLists.txt
@@ -15,6 +15,7 @@ set(UTILS_SOURCES
     grouppers/generator_utils.hpp
     grouppers/hdr_generator.cpp
     grouppers/hdr_generator.hpp
+    implementation/faces_saver.cpp
     batch_face_detector.cpp
     batch_face_detector.hpp
     collection_scanner.cpp
